+++ patches/src/main/kotlin/app/revanced/patches/shared/misc/audio/Fingerprints.kt
@@ -1,7 +1,7 @@
 package app.revanced.patches.shared.misc.audio
 
 import app.revanced.patcher.fingerprint
-import app.revanced.util.containsLiteralInstruction
+import app.revanced.util.literal
 import com.android.tools.smali.dexlib2.AccessFlags
 
 internal val formatStreamModelToStringFingerprint = fingerprint {
@@ -20,10 +20,7 @@ internal val formatStreamModelToStringFingerprint = fingerprint {
 internal const val AUDIO_STREAM_IGNORE_DEFAULT_FEATURE_FLAG = 45666189L
 
 internal val selectAudioStreamFingerprint = fingerprint {
-    accessFlags(AccessFlags.PUBLIC, AccessFlags.STATIC)
-    returns("L")
-    custom { method, _ ->
-        method.parameters.size > 2 // Method has a large number of parameters and may change.
-                && method.containsLiteralInstruction(AUDIO_STREAM_IGNORE_DEFAULT_FEATURE_FLAG)
+    literal {
+        AUDIO_STREAM_IGNORE_DEFAULT_FEATURE_FLAG
     }
 }